booth_manager/requirements.txt
PySide6>=6.5.0
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
PyMuPDF>=1.23.0
rarfile>=4.0
